Add files via upload
This is this starting of the layout of Conway's Life game.  I'm trying to make it so that having different states is as easy as just adding them to State.cs and that any number of dimensions should be handled quickly.  Visualizing higher dimensional life games might be tricky, but this should be able to determine it.